Added file. Added prompt for new parameter.
Added a jupyter notebook file for demonstrating model_output_manager.
Modified model_output_manager to prompt the user when a new parameter
is passed in the parameter dictionary that isn't yet present in
the parameter table. This prompt request a value to use to set
the previous values of this parameter in the parameter table.