Totally rewrote model_output_manager.py.
Totally rewrote model_output_manager.py to be much simpler. These
changes take away some of the functionality of MOM that wasn't
actually very helpful, like the method for saving output.
Now there are two main methods:
  1. run_exists, which checks to see if a run matching the provided
  parameters exists
  2. get_run_entry, which returns a unique id for a run with the
  provided parameters.
Creating the directory for output and writing the output is
now left to the user, since this is very straightforward given
the unique id for a run and since there are a number of different ways
the user may want to go about doing this.
There is also a Memory class that resembles the joblib Memory class,
with less functionality. Two major differences between joblib Memory
and MOM Memory are that:
  1. MOM creates a csv table documenting the runs.
  2. MOM does not invalidate the cache if the memoized function changes.
Note also that MOM can expand the number of parameters used to specify
runs without invalidating the cache. This is accomplished by asking the
user to specify a value for the parameter for the runs where the
parameter was not previously provided.