Added features to Memoize and changed output.
Added the feature to specify ignored parameters in Memoize.cache.
Also in Memoize.cache, changed the saved output behavior to create
a subdirectory with the name of the function being memoized and
to not create a subdirectory for each individual run, instead
using different filenames in the same directory.

Added licensing statements as a consequence of incorporating more
code from joblib.